requirements.txt
selenium>=4.0.0
rich>=13.0.0
chess>=1.9.0
python-chess>=1.0.0
pyyaml>=6.0

# pip install -r requirements.txt